# Improve EAD import performance for unnumbered components