I noticed that importing certain EAD files would take a long time.  Two in particular, both exported from Smithsonian's ArchivesSpace instance, ran for over an hour before I gave up.

The thing common to the slow files was that both used unnumbered component tags (i.e. `<c>` instead of `<c01>`).  The profiler showed all of the CPU time in this method:

  https://github.com/archivesspace/archivesspace/blob/3663176be821028007b5cdae2c5b01b4ea550e21/backend/app/converters/ead_converter.rb#L28-L38

which is the overridden version of this one:

  https://github.com/archivesspace/archivesspace/blob/3663176be821028007b5cdae2c5b01b4ea550e21/backend/app/converters/lib/xml_sax.rb#L124-L136

the EAD converter version has been modified to short-circuit the check for common tags--including numbered components--but still checks unnumbered components.  So that explains why I'm seeing different performance between my test files.

The trouble with all of this is that the importer is trying to work out whether a given node is empty, but the SAX parsing makes that hard.  Calling `node.inner_xml.strip.empty?` gives the right result, but slows down exponentially as every node in the document traverses every node under it, serializing them all into one huge string and then cloning that string with the final call to `.strip`.

Rather than adding more exceptions, I've worked on speeding up the "is empty?" check.  It requires two passes over the input XML (one to determine which nodes are empty, then one to do the original work), but the end result is much faster.  Here are some indicative numbers:

| File           | Size  | Original version | This version   |
|----------------|-------|------------------|----------------|
| copland.xml    | 5.8MB | 1 min 45 sec     | 1 min 40 secs  |
| Smithsonian 1  | 87MB  | never finished   | 18 mins        |
| hornstein.xml  | 468KB | 25 secs          | 15 secs        |
| Smithsonian 2  | 10MB  | never finished   | 2 mins 10 secs |
| Smithsonian 3  | 3KB   | 5 secs           | 5 secs         |

For the ones that finished, I've verified the results by re-exporting them as EAD under both versions of the code and comparing the results.